+++ lib/legion/extensions/microsoft_teams.rb
@@ -19,6 +19,8 @@ require 'legion/extensions/microsoft_teams/runners/cache_ingest'
 require 'legion/extensions/microsoft_teams/helpers/high_water_mark'
 require 'legion/extensions/microsoft_teams/helpers/prompt_resolver'
 require 'legion/extensions/microsoft_teams/helpers/session_manager'
+require 'legion/extensions/microsoft_teams/helpers/subscription_registry'
+require 'legion/extensions/microsoft_teams/helpers/token_cache'
 
 # Transport
 if defined?(Legion::Transport)